requirements.txt
openai
pydantic
langchain
langchain-openai
langchain-core
langgraph
python-dotenv
pandas
numpy
matplotlib
seaborn
tqdm